+++ build-standard-locally.sh
@@ -3,5 +3,5 @@ rm -rf Config
 mkdir Config
 go run .
 cp -r Config-Shared/* Config
-cp -r Config-CrystalHell/* Config
+cp -r Config-Standard/* Config