+++ gen/progression-standard.go
@@ -6,21 +6,21 @@ import (
 	"strings"
 )
 
-// VanillaProgression is responsible for producing content for progression.xml
-type VanillaProgression struct{}
+// StandardProgression is responsible for producing content for progression.xml
+type StandardProgression struct{}
 
 // GetPath returns file path for this producer
-func (*VanillaProgression) GetPath() string {
-	return "Config-Vanilla"
+func (*StandardProgression) GetPath() string {
+	return "Config-Standard"
 }
 
 // GetFilename returns filename for this producer
-func (*VanillaProgression) GetFilename() string {
+func (*StandardProgression) GetFilename() string {
 	return "progression.xml"
 }
 
 // Produce xml data to the provided channel
-func (p *VanillaProgression) Produce(c chan string) {
+func (p *StandardProgression) Produce(c chan string) {
 	defer close(c)
 	c <- `<config>`
 	c <- `<set xpath="/progression/perks/perk[@name='perkLivingOffTheLand']/@max_level">5</set>`
@@ -36,31 +36,31 @@ func (p *VanillaProgression) Produce(c chan string) {
 	c <- `</config>`
 }
 
-func (*VanillaProgression) getTraitTagsEnhanced() (tags []string) {
+func (*StandardProgression) getTraitTagsEnhanced() (tags []string) {
 	for _, plant := range data.Plants {
-		tags = append(tags, plant.GetSchematicName(""))
+		tags = append(tags, fmt.Sprintf("planted%s1_", plant.GetName()))
 	}
 	return
 }
 
-func (*VanillaProgression) getTraitTagsSingles() (tags []string) {
+func (*StandardProgression) getTraitTagsSingles() (tags []string) {
 	for _, plant := range data.Plants {
 		for _, trait := range data.Traits {
 			if plant.IsCompatibleWith(trait) {
-				tags = append(tags, plant.GetSchematicName(string(trait.Code)))
+				tags = append(tags, fmt.Sprintf("planted%s1_%c", plant.GetName(), trait.Code))
 			}
 		}
 	}
 	return
 }
 
-func (*VanillaProgression) getTraitTagsDoubles() (tags []string) {
+func (*StandardProgression) getTraitTagsDoubles() (tags []string) {
 	for _, plant := range data.Plants {
 		for _, first := range data.Traits {
 			if plant.IsCompatibleWith(first) {
 				for _, second := range data.Traits {
 					if first.IsCompatibleWith(second) && plant.IsCompatibleWith(second) {
-						tags = append(tags, plant.GetSchematicName(string(first.Code)+string(second.Code)))
+						tags = append(tags, fmt.Sprintf("planted%s1_%c%c", plant.GetName(), first.Code, second.Code))
 					}
 				}
 			}