+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - remove seed return chance during harvest
 - add recipe for enhanced seed crafting by hand
+- rename vanilla->researcher, crystal-hell->standard
 
 ## [0.6.0] - 2022-05-19